Refactoring SearchFilter.java into multiple SearchFilters
