+++ src/service/IncentiveSearchFilter.java
@@ -1,0 +1,13 @@
+package service;
+
+public class IncentiveSearchFilter extends SearchFilter {
+  /*
+  SearchFilter implementation for Incentive GUI
+ */
+  int dealerID;
+
+  public IncentiveSearchFilter(int dealerID) {
+    super();
+    this.dealerID = dealerID;
+  }
+}
